refactor: centralize logger
