+++ hooks/useDailyQuote.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { Alert } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { DAILY_QUOTE_ENABLED } from '@/constants/featureFlags';
-import * as logger from '@/helpers/logger';
+import * as logger from '@/shared/logger';
 
 const QUOTES = [
   'Believe in yourself!',
