+++ shared/logger.js
