+++ tests/setupLogger.ts
@@ -1,4 +1,4 @@
-jest.mock('@/helpers/logger', () => ({
+jest.mock('@/shared/logger', () => ({
   log: jest.fn(),
   warn: jest.fn(),
   error: jest.fn(),
